requirements.txt
pygame
requests
python-dotenv
python-socketio[client]
websocket-client